base-requirements.txt
pyspark==3.5.5
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-http